requirements.txt
langchain-openai
pydantic~=2.11.10
python-dotenv~=1.1.1
pytest~=8.4.2
langchain~=0.3.27